FinAssist/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
openai
faiss-cpu
scikit-learn
pandas
numpy
python-multipart
python-dotenv
statsmodels
requests
